requirements.txt
Flask
qrcode
google-cloud-datastore
google-cloud-core
google-cloud-storage